WIP - Figuring out how to check for changes in current document prior to exiting.
Changes: Encapsulated what I was trying to do to detect changes into a method that CodeEditor can invoke. Tweaked the FileWatcher class and I believe it can now determine when a change has been made to the current file. However, WatchService is apparently very slow when used on a Mac (what I have). Needs further testing on a non-Mac to confirm if this is an avenue worth pursuing.