WIP - Figuring out how to check for changes in current document prior to exiting.
Changes: Found a much simpler way to get a snapshot of what is in a document when it's opened and right before someone tries to close it. Still need to figure out how to add a pop-up warning with a button that allows user to save when prompted, then exit.